@@ -1,12 +1,11 @@
 from __future__ import annotations
 
+from collections.abc import Mapping
 from contextlib import suppress
 from pathlib import Path
 from typing import TYPE_CHECKING
 from typing import Any
 from typing import ClassVar
-from typing import Dict
-from typing import Mapping
 from typing import Union
 
 from cleo.helpers import option
@@ -27,7 +26,7 @@ if TYPE_CHECKING:
 
     from poetry.repositories import RepositoryPool
 
-Requirements = Dict[str, Union[str, Mapping[str, Any]]]
+Requirements = dict[str, Union[str, Mapping[str, Any]]]
 
 
 class InitCommand(Command):
