@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 from typing import TYPE_CHECKING
-from typing import List
 
 from poetry.packages.dependency_package import DependencyPackage
 
@@ -13,7 +12,7 @@ if TYPE_CHECKING:
     from poetry.core.packages.package import Package
 
 
-class PackageCollection(List[DependencyPackage]):
+class PackageCollection(list[DependencyPackage]):
     def __init__(
         self,
         dependency: Dependency,
