@@ -7,7 +7,6 @@ from cleo.helpers import argument
 from cleo.helpers import option
 
 from poetry.console.commands.command import Command
-from poetry.utils._compat import is_relative_to
 
 
 if TYPE_CHECKING:
@@ -55,8 +54,8 @@ class EnvRemoveCommand(Command):
             self.line(f"Deleted virtualenv: <comment>{venv.path}</comment>")
         if remove_all_envs or is_in_project:
             for venv in manager.list():
-                if not is_in_project or is_relative_to(
-                    venv.path, self.poetry.pyproject_path.parent
+                if not is_in_project or venv.path.is_relative_to(
+                    self.poetry.pyproject_path.parent
                 ):
                     manager.remove_venv(venv.path)
                     self.line(f"Deleted virtualenv: <comment>{venv.path}</comment>")
