@@ -7,7 +7,6 @@ import shutil
 
 from pathlib import Path
 from typing import TYPE_CHECKING
-from typing import Iterator
 
 import pytest
 
@@ -29,6 +28,8 @@ from poetry.utils.env import ephemeral_environment
 
 
 if TYPE_CHECKING:
+    from collections.abc import Iterator
+
     from pytest_mock import MockerFixture
 
     from poetry.poetry import Poetry
